feat: Generate OpenApi spec during build
