+++ platform/src/apis/Platform.Api.ChartRendering/src/api.ts
@@ -1,0 +1,2 @@
+export { GetHealthApi } from "./functions/healthCheck/api";
+export { GetVerticalBarChartApi } from "./functions/verticalBarChart/api";
