+++ platform/src/apis/Platform.Api.ChartRendering/src/functions/healthCheck/api.ts
@@ -1,0 +1,12 @@
+import { ApiMapper } from "ts-oas";
+
+export type GetHealthApi = ApiMapper<{
+  path: "/health";
+  method: "GET";
+  responses: {
+    /**
+     * @contentType text/plain
+     */
+    "200": string;
+  };
+}>;
