+++ platform/src/apis/Platform.Api.ChartRendering/src/functions/healthCheck/route.ts
@@ -1,0 +1,8 @@
+import { app } from "@azure/functions";
+import { healthCheck } from "./healthCheck";
+
+app.http("health", {
+  methods: ["GET"],
+  authLevel: "anonymous",
+  handler: healthCheck,
+});
