+++ platform/src/apis/Platform.Api.ChartRendering/src/functions/index.d.ts
@@ -24,8 +24,3 @@ type ChartBuilderResult = {
   id: string;
   html: string | undefined;
 };
-
-type ChartJsBuilderResult = {
-  id: string;
-  buffer: Buffer<ArrayBufferLike>;
-};
