+++ platform/src/apis/Platform.Api.ChartRendering/src/functions/verticalBarChart/builder.ts
