+++ platform/src/apis/Platform.Api.ChartRendering/src/functions/verticalBarChart/route.ts
@@ -1,0 +1,8 @@
+import { app } from "@azure/functions";
+import { verticalBarChart } from "./function";
+
+app.http("verticalBarChart", {
+  methods: ["POST"],
+  authLevel: "anonymous",
+  handler: verticalBarChart,
+});
