+++ platform/src/apis/Platform.Api.ChartRendering/src/functions/verticalBarChart/validator.ts
