+++ platform/src/apis/Platform.Api.ChartRendering/src/index.ts
@@ -1,26 +1,12 @@
 import appInsights from "applicationinsights";
 import { app } from "@azure/functions";
-import { healthCheck } from "./functions/healthCheck/healthCheck";
-import { verticalBarChart } from "./functions/verticalBarChart/verticalBarChart";
 
 appInsights
   .setup()
   .setAutoCollectRequests(true)
-  .setSendLiveMetrics(true)
+  .setSendLiveMetrics(false)
   .start();
 
 app.setup({
   enableHttpStream: true,
-});
-
-app.http("health", {
-  methods: ["GET"],
-  authLevel: "anonymous",
-  handler: healthCheck,
-});
-
-app.http("verticalBarChart", {
-  methods: ["POST"],
-  authLevel: "anonymous",
-  handler: verticalBarChart,
 });
