+++ platform/src/apis/Platform.Api.ChartRendering/tsconfig.json
@@ -104,7 +104,8 @@
   },
   "exclude": [
     "node_modules",
-    "tests"
+    "tests",
+    "**/api.ts"
   ],
   "tsc-alias": {
     "resolveFullPaths": true,
